feat: add SSO definitions
Small String Optimization is a common trick used by `std::string`
implementations in C++. Instead of allocating every string on the
heap, we can reuse the `pstring_t` structure itself to store small
strings on the stack. This reduces both the memory overhead of
the structure and the allocation speed.

On 64-bit architectures, `pstring_t` can store up to 22 characters
of the 32 bytes allocated. While there are tricks to achieve much
higher ratios of used bytes (up to 31!), they are usually less
maintainable and portable. The current design allows branchless
reading of string's length and capacity, as well as a direct
access to the character buffer. Unlike many other implementations,
this SSO design avoids endian-dependent tricks at all costs.